pyproject.toml
[project]
name = "quizy-app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
